refactor/ Response 객체를 사용하도록 API 수정
